final_submit/hashtag_rd1/prediction/requirements.txt
click
h11
pydantic
starlette
fastapi
openai
uvicorn
python-dotenv
pandas